new iterations as requested
notes from darcy:

Here are just a few things I was hoping could be tweaked:
Can the slider on the home page rotate through the images faster?
Maybe 5 seconds per image? Also is there a way that the next image can be ‘loaded’ faster so it doesn’t slide to a grey box before the image appears? - Give me a ring if this isn’t clear haha.
On the projects page can we replace the I Am Horror tile with the image attached? Let me know if it doesn't work or if you need a better file quality (and then put the image of the lady screaming on the I Am Horror project page)
Can we swap the order of Dimboola and Pop Portrait projects?
For the credits of each project can we reduce the text size and make it a lighter grey?
Can you please crop the image of The Removalists so there isn't a black bar on the bottom of it.
Can you add an icon link to my instagram account - maybe just on the contact me page?
Can you please update the about me page with the text attached :)